backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.21
psycopg2-binary>=2.9.0
alembic==1.13.1
pydantic-settings==2.3.4
httpx==0.27.0
loguru==0.7.2
python-dotenv==1.0.1
numpy>=1.26.4
scipy>=1.13.0
skyfield==1.49
astropy>=6.1.0
pandas>=2.0.0
# gmn-python-api==0.0.13 (commented out due to pandas 1.3.5 dependency conflict on Python 3.14)
requests==2.32.3
nrlmsise00==0.1.1
celery==5.4.0
redis==5.0.6
flower==2.0.1
apscheduler==3.10.4
beautifulsoup4==4.12.3
lxml==5.2.1
